requirements.txt
asgiref==3.9.2
beautifulsoup4==4.14.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Django==5.2.6
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
FlightRadarAPI==1.4.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
Markdown==3.9
packaging==25.0
psycopg==3.2.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
